TEMPLATE_PROJECT/requirements.txt
# WidgetTDC Platform Template - Python Dependencies

# Core orchestration
pyyaml>=6.0

# LLM Integration
anthropic>=0.25.0

# Optional: Advanced orchestration (LangChain)
# Uncomment if using LangChain features
# langchain>=0.1.0
# langchain-anthropic>=0.1.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Monitoring & Logging
python-dotenv>=1.0.0
